+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0.0
 
+-   Add max_uses to templates
 -   Don't make slugs unique
 -   Throw an exception when saving an entity with a non-unique path
 -   Soft delete paths for soft deleted entities